### 2020-04-17 #28 @aleblois
[Re] Reproduction of the normal and pathological dynamics in the basal ganglia-thalamo-cortical network
**Original article**:  Leblois A, Boraud T, Meissner W, Bergmann H, Hansel D. Competition
between feedback loops underlies normal and pathological dynamics in the
basal ganglia. J Neurosci 26: 3567–3583, 2006.

**PDF URL**: https://github.com/aleblois/BGdynamics2006/blob/master/ReScience_C__BGdynamics2006.pdf
**Metadata URL**: https://github.com/aleblois/BGdynamics2006/blob/master/metadata.tex
**Code URL**: https://github.com/aleblois/BGdynamics2006
**yaml file**: https://github.com/aleblois/BGdynamics2006/blob/master/metadata.yaml

**Scientific domain: Neuroscience**
**Programming language: C and Matlab**
**Suggested editor**: Etienne Roesch, Tiziano Zito, Benoît Girard, Nicolas Rougier
