+++ metadata.tex
@@ -1,0 +1,43 @@
+% DO NOT EDIT - automatically generated from metadata.yaml
+
+\def \codeURL{https://github.com/aleblois/BGdynamics2006}
+\def \codeDOI{}
+\def \dataURL{}
+\def \dataDOI{}
+\def \editorNAME{}
+\def \editorORCID{}
+\def \reviewerINAME{}
+\def \reviewerIORCID{}
+\def \reviewerIINAME{}
+\def \reviewerIIORCID{}
+\def \dateRECEIVED{17 April 2020}
+\def \dateACCEPTED{}
+\def \datePUBLISHED{}
+\def \articleTITLE{[Re] Reproduction of the normal and pathological dynamics in the basal ganglia-thalamo-cortical network}
+\def \articleTYPE{Reproduction}
+\def \articleDOMAIN{}
+\def \articleBIBLIOGRAPHY{bibliography.bib}
+\def \articleYEAR{2020}
+\def \reviewURL{}
+\def \articleABSTRACT{In 2006, during my PhD, I published my first modelling paper showing that the normal and pathological dynamics of the basal ganglia-thalamo-cortical network could be understood from a theoretical standpoint relying on a simplified model of the circuit. I recently was told by a colleague about a challenge to reproduce modelling studies over 10-years old. I was quite confident I could reproduce the results of my 2006 paper, which was based on a combination of analytical calculation and computer simulations programmed in C with figures produced through Matlab custom code. All the code was still saved on my computer (although in a simple but relatively messy file-and-folder organisation), and I went through my personal archives and re-run ancient code until I could provide again the results described here. }
+\def \replicationCITE{}
+\def \replicationBIB{}
+\def \replicationURL{}
+\def \replicationDOI{}
+\def \contactNAME{Arthur Leblois}
+\def \contactEMAIL{arthur.leblois@u-bordeaux.fr}
+\def \articleKEYWORDS{Reproduction, Neuroscience, Modelling, Parkinson’s disease, Oscillations}
+\def \journalNAME{ReScience C}
+\def \journalVOLUME{4}
+\def \journalISSUE{1}
+\def \articleNUMBER{}
+\def \articleDOI{}
+\def \authorsFULL{Arthur Leblois}
+\def \authorsABBRV{A. Leblois}
+\def \authorsSHORT{Leblois}
+\title{\articleTITLE}
+\date{}
+\author[1,\orcid{0000-0002-9392-5939}]{Arthur Leblois}
+
+\affil[1]{Institut des Maladies Neurodégénératives, UMR CNRS 5293, Neurocampus, Université de Bordeaux, 146 rue Léo Saignat, 33000 Bordeaux, France.}
+
