+++ simdelai2boucles_all2all_fixedepswitch_tpsyndif.c
@@ -1,0 +1,184 @@
+#include  <stdio.h>
+#include <math.h>
+/*#include <unistd.h>*/
+#include <fcntl.h>
+
+double g(double x){
+	if (x>0) return x;
+	else return 0;
+}
+double sign(double x){
+	return x/fabs(x);
+}
+main(){
+	int i, j, k, m, Dstrctx, Dgpistr, Dthagpi, Dctxtha, Dstnctx, Dgpistn;
+	double a, error, t, tf, to, temp, itempo, eps, Gm, Gp, control1, control2, control3, ictxmin, mstr0, mstn0, mstr1, mstn1, mtha10, mtha0, mtha1, mgpi0, mgpi10, mgpi1,mctx1, Gstrctx0, Ggpistr0, Gthagpi0, Gctxtha0, Gstnctx0, Ggpistn0, Gstr10, Ggpi10, Gctx10, istr, igpi, ictx, ithal, Gstrctxg0, Ggpistrg0, Gthagpig0, Gctxthag0, Gstnctxg0, Ggpistng0, Gstrg10, Ggpig10, Gctxg10, istrg, igpig, ictxg, ithalg, fstrg[20], fgpig[20], fctxg1[20], fctxg2[20], fthag[20], dt, pi, isstrg, isgpig, isctxg, isthalg, Gstrctxg, Gstrgl, Ggpistrg, Ggpigl, Gthagpig, Gctxthag, Gstnctxg, Ggpistng, Gctxgl, isctxextg, isstn, istn, isstng, istng, fstng[20], itemp, Gstrctxd0, Ggpistrd0, Gthagpid0, Gctxthad0, Gstnctxd0, Ggpistnd0, Gstrd10, Ggpid10, Gctxd10, istrd, igpid, ictxd, ithald, fstrd[20], fgpid[20], fctxd1[20], fctxd2[20], fthad[20], isstrd, isgpid, isctxd, isthald, Gstrctxd, Gstrdl, Ggpistrd, Ggpidl, Gthagpid, Gctxthad, Gctxdl, isctxextd, isstnd, istnd, fstnd[20], Gstnctxd, Ggpistnd, gamma;
+	FILE *tab, *tubd, *tubg;
+	tubd=fopen("evold.tab","w");
+	tubg=fopen("evolg.tab","w");
+	tab=fopen("tempscvg.tab","a");
+
+	pi=3.14159265; dt=.1; tf=1000; t=0; temp=0; to=-1;
+	Dstrctx=12; Dgpistr=19; Dthagpi=10; Dctxtha=10;Dstnctx=10;Dgpistn=10;	/*Valeur des delais*/
+	gamma=0.3; eps=0.3; a=1.85;		/*Dopamine, couplage et diff d'input*/
+
+	/*On fixe les valeurs des activités dans 0/0 et 0/1:*/
+	mstr0=1; mstn0=20; mstr1=15; mstn1=60;	mctx1=20;
+	mtha0=20; mtha1=30; mgpi0=70; mgpi10=125; mgpi1=40;
+
+	fprintf(tab,"%f %f %f %f %f %f %f %f %f %f %f\n",mstr0,mstr1,mgpi0,mgpi1,mctx1,mtha0,mtha1,mstn0,mstn1,eps,gamma);
+
[…]
