+++ simdelai2boucles_random_GPe_Dopa_str_as2006.c
@@ -1,0 +1,949 @@
+#include  <stdio.h>
+#include <math.h>
+/*#include <unistd.h>*/
+#include <fcntl.h>
+#include <stdlib.h>
+#include "random.h"
+
+double g(double x){
+	if (x>0) return x;
+	else return 0;
+}
+double sign(double x){
+	return x/fabs(x);
+}
+int main(){
+	int n, n1, nmoy, nunit, nctx, nstr, ngpi, ntha, nstn, ngpe, i, j, k, D, Dstrctx, Dgpistr, Dthagpi,
+		Dctxtha, Dgpigpe, Dgpestr, Dgpestn, Dstngpe, Dstnctx, Dgpistn, step;
+	long u;
+
+	double t, tf, tf1, tf2, dt, mvt_dur, trampe, t_str, t_mvt, pi, eps, Dopa, pcon[15], sigma[6], tpsmoy, temp1, mctx1,
+		temp, gamma, broad, pic, tobruit, I_ctx_mvt,
+		Gstrctx0,	Ggpistr0, Gthagpi0, Gctxtha0, Gstnctx0, Ggpistn0, Gstr10, Ggpi10, Gctx10,
+		Gstrctxg0, Ggpistrg0, Gthagpig0, Gctxthag0, Ggpigpeg0,
+		Ggpestrg0, Ggpestng0, Gstngpeg0, Gstnctxg0, Ggpistng0, Gstrg10, Ggpig10, Gctxg10,
+		ictx0, istr0, igpi0, igpe0, istn0, itha0, I_select_str,
+		moyctxg, moystrg, moygpig, moystng, moythag, moygpeg,
+		fmoystrg, fmoystng, fmoythag, fmoygpeg,
+		Gstrctxd0, Ggpistrd0, Gthagpid0, Gctxthad0, Ggpigped0, Ggpestrd0, Ggpestnd0,
+		Gstngped0, Gstnctxd0, Ggpistnd0, Gstrd10, Ggpid10, Gctxd10,
+		moyctxd, moystrd, moygpid, moystnd, moythad, moygped,
+		fmoystrd, fmoystnd, fmoythad, fmoygped,
+		n_connect_ctx, n_connect_stn;
+
+	double* v1 = (double*)malloc(1*sizeof(double));
+	double* v3 = (double*)malloc(1*sizeof(double));
+	double* fmoygpig = (double*)malloc(400*sizeof(double));
+	double* fmoygpid = (double*)malloc(400*sizeof(double));
+	double* fmoyctxg = (double*)malloc(400*sizeof(double));
+	double* fmoyctxd = (double*)malloc(400*sizeof(double));
+	double* bgctxg = (double*)malloc(400*sizeof(double));
+	double* bgctxd = (double*)malloc(400*sizeof(double));
+	double* reststrg = (double*)malloc(400*sizeof(double));
+	double* reststng = (double*)malloc(400*sizeof(double));
+	double* restgpig = (double*)malloc(400*sizeof(double));
+	double* restgpeg = (double*)malloc(400*sizeof(double));
+	double* restctxg = (double*)malloc(400*sizeof(double));
+	double* restthag = (double*)malloc(400*sizeof(double));
+	double* istr = (double*)malloc(400*sizeof(double));
+	double* istn = (double*)malloc(400*sizeof(double));
+	double* igpi = (double*)malloc(400*sizeof(double));
+	double* igpe = (double*)malloc(400*sizeof(double));
+	double* icortex = (double*)malloc(400*sizeof(double));
+	double* ictx = (double*)malloc(400*sizeof(double));
+	double* ithal = (double*)malloc(400*sizeof(double));
+	double* istrg = (double*)malloc(400*sizeof(double));
[…]
