### 2020-04-27 #28 @rougier
@benoit-girard Can you take care of this submission (either assign an editor or you can edit it youself). Only one reviewer needed since this is for the Ten Years Reproducibility Challenge.
