### 2020-05-26 #28 @benoit-girard
I think I can review it myself.
