### 2020-06-15 #28 @aleblois
Salut,On discute 5 min avant le début du cours?Arthur
    Le lundi 15 juin 2020 à 07:43:17 UTC+2, Nicolas P. Rougier <notifications@github.com> a écrit :

Gentle reminder
