### 2020-06-22 #28 @benoit-girard
Overall, the paper is an interesting contribution to the 10-year reproducibility challenge. I suggest the correction of a number of minor points listed below:

Page 3, you mention that you had to correct a few mistakes in the matlab code to reproduce the analyses and the figures. Could you be more specific, as the paper is precisely about reproduction: what sort of mistakes are you talking about? Is it that the latest version of matlab required adjustments of your old code because of compatibility problems? Or that you found only on old version of the code that wouldn't have worked even in 2006, and that you had to re-make corrections you already did back in that time?

Is the figure 2 in the current manuscript a new figure generated in 2020, or a copy of the original figure? The compression artifacts tend to suggest it is the latter hypothesis, but it is not absolutely clear.

Page 2, you write "The description of the network studied in the 2006 paper (Figs 1&2 and Methods)", it is not clear if you are talking about the figures and methods of the 2006 paper, or of the current paper. As Fig. 2 of the curent paper does not really describe the model, and as there is no method section in the current paper, I assume you are talking about the 2006 one, but you should then state that explicitly.

Page 5, you write that you were able to produce "a set of figures identical or similar to the figures of the original paper"; are you referring to figures 3, 4 and 5? If yes, could you state it explicitly?

The following sentence in the introduction probably require reformulation: "The model for understanding how BG dysfunction leads to motor symptoms in (PD) and Huntingtonʼs disease was proposed 1,2" did you mean something like "The classical model for understanding how BG dysfunction leads to motor symptoms in (PD) and Huntingtonʼs disease has been proposed by 1,2"?

Formatting problems:
* page 2, the reference to Leblois et al. 2006 was not correctly compiled in "we designed a theoretical model of the BG-thalamo-cortical loop with a simplified connectivity pattern that did not require a segregation between the direct and indirect pathwaysLeblois:2006"
* Two sentences lost in the middle of the text seem to be incorrectly formatted subsection titles: "Description of the network model" (page 2) and "Retrieval of the software" (page 3).
* page 3: you have two dots at the end of  "the phase diagram(Fig. 2).."
* page 5: a number of references to figures were not compiled correctly, you end up with "??" rather than figure numbers on 7 occasions.
* page 5: you wrote "weekly", did you mean "weakly"?
* page 5, in the sentence "Rather, reproduction of the results consist in finding reproducing the behaviour of the network": "consist"->"consists", "finding reproducing" choose one or the other.
* page 5: "neural populations pf the network are activated" -> of the network
* page 7: "It provides a global understanding of the network dynamics the guide the numerical exploration" -> "that guides the numerical exploration" ?
