### 2020-06-25 #28 @aleblois
Thank you for your comments. Here are my point-by-point answers, I hope they address the issues you raised.

Page 3, I have corrected and added detail to explain what had to be changed in my previous matlab code. The text I added to the paper is the following: " First, a few functions had changed names or syntax and I had to correct the command line where I called the function to make my scripts compatible with the current version of Matlab I use. Secondly, the convention on the parameters and their values given in the text were not consistent between my Matlab code and the published paper. Indeed, a late change in a convention concerning one parameter (the cross-circuit weight gamma) was not defined in the same way in the published article and in most of the work I did before. We had changed this convention very late in the writing of the paper, and I had taken no note on this parameter change. In the published paper, the ratio between the STN-GPi connection gain within a loop and across loops was gamma, while in the previous paper it was gamma/(1-gamma). Once this parameter change was taken into account, all results could be retrieved. Archiving all conventions used for all parameters of the model is crucial for future use. "

Figure 2 was rebuild, as all other figures, and is not a copy from the original (the axes limits are actually slightly different). The compression artifact is simply due to the jpg export of the matlab original figure.

In the description of the network model, I now explicitely refer to Fig1&2 and methods of the original 2006 paper.

Page 5, I now refer explicitely to figures 3-5.

The introduction sentence is now corrected.

All formatting problems have been corrected. Thanks again for the careful editing, and my apologies for the many misprints and formatting errors.
