### 2020-06-25 #28 @aleblois
Revised version:

PDF URL: https://github.com/aleblois/BGdynamics2006/blob/master/ReScience_Leblois2006_revisedversion.pdf
Metadata URL: https://github.com/aleblois/BGdynamics2006/blob/master/metadata.tex
Code URL: https://github.com/aleblois/BGdynamics2006
