### 2020-06-30 #28 @benoit-girard
I am fully satisfied with the revised version of the paper, I recommend publication.
