### 2020-06-30 #28 @benoit-girard
@rougier will you take care of the publication process? If not, is it still the same old long process, or has it changed somehow?
