### 2020-06-30 #28 @rougier
There's a new publication process that is less painful but not yet fully automatic. Good news is that you can first test if on the Zenodo sandbox server and if everything works ok, you can do the actual publication. Else, I can do it, just tell me.
