### 2020-09-08 #28 @rougier
Sorry for the delay. There's a new publication process available at  https://github.com/ReScience/articles which is supposed to be smoother. Do not forget to use the sandbox to check everything is fine first.
