### 2020-09-28 #28 @benoit-girard
@rougier first step, first questions...
* there is a metadata.tex file in @aleblois 's repository, but I can't find the yaml one. Am I missing something? Should @aleblois provide it?
* the pdf currently does not indicate the name of the editor, of the reviewer, and the date of acceptance. Is it still OK for the first steps of the submission? Do I need to edit that myself (but I do not have the source, only the pdf...), or ask @aleblois to do it?
