### 2020-09-29 #28 @benoit-girard
OK, reading the latest instructions for authors, I see that the author should provide the "mandatory" yaml file.
@aleblois can you do that please?
