### 2020-10-14 #28 @aleblois
Actually, the yaml file seems o have the wrong info (Rougier and Hinsen paper). I'm not sure how to edit/generate the yaml. Nicolas, any tip on this?
