### 2020-10-14 #28 @aleblois
I edited manually the yaml file, which should be OK now. Best
