### 2020-10-15 #28 @rougier
@aleblois Actually, this is the way to got with the yaml file (manual editing).
@benoit-girard tell me if you need help with the publication process.
