### 2021-02-16 #28 @benoit-girard
OK, i am (finally) in the publication process. To which volume/issue should I associate the paper? Volume 6, Issue 1, Number 19?
