### 2021-02-16 #28 @benoit-girard
@aleblois I have send a pull request on your repo, with an updated metadata.yaml file.
If I am not mistaken (@rougier ?), you will have to validate the request, and update your article.pdf with the new information (volume, issue, number, doi, editor, reviewer, dates of acceptance & publication).
