### 2021-02-16 #28 @benoit-girard
This : https://github.com/aleblois/BGdynamics2006/pull/1
