### 2021-02-16 #28 @aleblois
I merged it, is it OK now? Sorry my Github experience is near zero....
