### 2021-02-16 #28 @benoit-girard
That's perfect!

Now you have to regenerate you pdf, including the new metadata (that you will find in the metadata.yaml file) :
volume 6, issue 1, number 19, doi 10.5281/zenodo.4543356, editor (me, orcid 0000-0002-8117-7064), reviewer (me also), dates of acceptance (June 30, 2020) & publication (February 16, 2021).

Tell me when it's done!
