### 2021-02-17 #28 @aleblois
In aleblois/BGdynamics2006#1 (https://github.com/aleblois/BGdynamics2006/pull/1), the file ReScience_C__BGdynamics2006.pdf is the new PDF. Should I put it somewhere in https://github.com/ReScience/articles? Not sure where....
