### 2021-02-17 #28 @benoit-girard
@rougier I have the pdf, I am now at the publish step (step 11), I get an error I do not (yet?) understand:

```girard$ ./publish.py --sandbox --metadata metadata.yaml --pdf article.pdf
Uploading content to Zenodo... Traceback (most recent call last):
  File "./publish.py", line 204, in <module>
    upload_content(server, token, article_id, article_file)
  File "./publish.py", line 25, in upload_content
    raise IOError("%s: " % response.status_code + response.json()["message"])
OSError: 404: PID does not exist.```
