### 2021-02-18 #28 @rougier
Important: you need to do tests first on the sandbox server

Did you do the process.sh step and get a pre DOI ?
