### 2021-02-18 #28 @benoit-girard
By the way, the error above resulted from:

`./publish.py --sandbox --metadata metadata.yaml --pdf article.pdf`

It was on sandbox
