### 2021-02-19 #28 @rougier
Did you update the metadata.yaml with the provided DOI (and ideally rebuild the PDF) ?
