### 2021-02-19 #28 @benoit-girard
Possible problem: metadata was containing the Zenodo doi (not the sandbox one). To test that possibility, I replaced the doi in metadata.yaml with the one I obtained when I did the process.py on the sandbox, relaunched the publish.py on sandbox, and now it tells me "Metadata is newer than PDF, probably PDF needs to be rebuilt". OK great, that's normal, but in the end I don't know if this test on sandbox worked or not, and if not, grmbl.
