### 2021-02-19 #28 @benoit-girard
Ok, with a little touch, it worked...

Well, I am not sure it fully worked: I obtained the following (some part of the output message documented at step 11 is not there, but there is no obvious error)
```
girard$ ./publish.py --sandbox --metadata metadata.yaml --pdf article.pdf
Uploading content to Zenodo... done!
Updating metadata to Zenodo... done!
Publishing on Zenodo... done!
Entry is online at https://sandbox.zenodo.org/record/730471
```
