### 2021-02-22 #28 @rougier
Yes, I can see it! So now you can upload it on the actual Zenodo server (and you'll able to merge the new branch in this case). Note that one difficulty is that you need to produce a new PDF one you obtained the DOI such as to publish the paper with the right DOI.

In case of problems, you still have the option to modify the entry on the Zenodo website directly.
