### 2021-03-08 #28 @rougier
Just got the notification on Zenodo (to be included in the ReScience collection)!
