### 2021-03-08 #28 @benoit-girard
I managed to do step 13, but I am unsure about the PR : I could change the bib file and commit the change, so is a PR really necessary?
