### 2021-03-08 #28 @rougier
The PR is mostly to tell me the bib item can be integrated such that I can rebuild the website (don't try to do it yourself or you might risk to break it).
