### 2021-03-08 #28 @benoit-girard
The thing is that as the branch has been merged peacefully, I have nothing to include in a PR.
