### 2021-03-08 #28 @rougier
Oh sorry, I was referring to the modification you need t do to the website (= add the new bib titem to the published.bib file)
