### 2021-03-09 #28 @benoit-girard
I added the new bib item in the published.bib file : you can rebuild the website!
